Move loan API calls into dedicated module
