+++ src/api/apiData.js
@@ -53,58 +53,6 @@ export const verifyIdentity = async (identity, value) => {
   return response;
 };
 
-// apply for a loan
-export const applyForLoan = async (loanFormData) => {
-  const response = await axios.post("/loans/apply", loanFormData);
-  return response;
-};
-
-// update loan details pending confirmation
-export const updatePendingLoanDetails = async (loanFormData, loanID) => {
-  const response = await axios.patch(`/loans/${loanID}/update`, loanFormData);
-  return response;
-};
-
-// confirm loan application
-export const confirmLoanApplication = async (loan_id, password) => {
-  const response = await axios.post("/loans/confirm", {
-    loan_id: loan_id,
-    password: password,
-  });
-  return response;
-};
-
-// fetch loan history with filters
-export const fetchLoans = async (page = 1, limit = 10, filters = {}) => {
-  const { query, startDate, endDate, status } = filters;
-  let queryString = `page=${page}&limit=${limit}`;
-
-  if (query) {
-    // If query is numeric, search by loan_amount; else treat as loan_id
-    if (!isNaN(Number(query))) {
-      queryString += `&loan_amount=${query}`;
-    } else {
-      queryString += `&loan_id=${query}`;
-    }
-  }
-
-  if (startDate) {
-    queryString += `&start_date=${startDate}`;
-  }
-
-  if (endDate) {
-    queryString += `&end_date=${endDate}`;
-  }
-
-  if (status) {
-    queryString += `&status=${status}`;
-  }
-  
-  const response = await axios.get(`/loans/history?${queryString}`);
-  return response.data;
-};
-
-
 // update preferences (sms, email, language)
 export const updatePreferences = async (
   emailState,
@@ -185,19 +133,6 @@ export const markAllAsRead = async () => {
 export const deleteNotification = async (notification_id) => {
   const response = axios.delete(`/notifications/${notification_id}`);
   return response;
-};
-
-
-// get loan by ID
-export const getLoanDetails = async (loanID) => {
-  const response = await axios.get(`/loans/${loanID}`);
-  return response;
-};
-
-// get borrowing limit
-export const getBorrowingLimit = async () => {
-  const response = await axios.get("/loans/borrowing-limit");
-  return response.data;
 };
 
 // log user out
